services/recruiter_assistant/requirements.txt
langchain_openai 
langchain_core
python-dotenv
langchain_community
langserve
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub 